Add an ACL for FORCE_SSL situation.
When haproxy behind other load balance proxy,
such as Amazon ELB, Aliyun SLB. If you use ELB to terminate the SSL, and the haproxy listen on HTTP only, then it will raises the "ERR_TOO_MANY_REDIRECTS" error.

But the load balancer can pass a additional X-Forwarded-Proto header to haproxy. So, if we add an ACL to process the X-Forwarded-Proto header, then will solve the problem.